pyproject.toml
[tool.poetry]
name = "python-advanced"
version = "0.1.0"
description = ""
authors = ["Stanislav Osipov <stanislav.osipov89@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
fastapi = "^0.115.12"
uvicorn = "^0.34.3"
pydantic-settings = "^2.9.1"
pytest = "^8.4.0"
selene = "^1.0.2"
mimesis = "^18.0.0"
python-dotenv = "^1.1.0"
fastapi-pagination = "^0.13.2"
email-validator = "^2.2.0"
psycopg2-binary = "^2.9.10"
sqlalchemy = "^2.0.41"
sqlmodel = "^0.0.24"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pytest-sugar = "^1.0.0"
pylint = "^3.3.7"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
log_cli = false
log_cli_level = "INFO"
addopts = "--tb=short --strict-markers"

markers = [
    "smoke: Smoke tests for basic functionality",
    "auth: Authentication related tests",
    "crud: CRUD operation tests",
    "pagination: Pagination related tests",
    "slow: Tests that take longer to run"
]